Add files via upload
The graphical user interface (GUI) code has been developed to streamline the feedback process for assignments, aiming to automate and enhance the efficiency of the feedback mechanism. This innovative solution is designed to significantly decrease the workload of the backend team by expediting the feedback process, ultimately resulting in time savings and improved operational effectiveness.